fix(ci): 3 hallazgos mas de la corrida 26 (GRANT ad-hoc, CHECK drift, cast pgTAP)
Con la version de supabase-cli pineada, la corrida 26 llego a ejecutar los 16 tests
de verdad y la mayoria ya pasa. Quedaban 3 causas claras (mas 1 de logica real que
se deja para la proxima):

1. "permission denied for table comprobantes/empresas" (centros_costo,
   sync_uala_to_bancos). Estas 2 son las UNICAS tablas de toda la base que ninguna
   migration crea (ad-hoc puras, anteriores al sistema de migrations) — y como
   ninguna migration hace GRANT explicito sobre NADA, el resto de las tablas
   depende de un mecanismo de Supabase Cloud (auto-otorga privilegios a
   anon/authenticated/service_role al crear una tabla) que un `supabase start`
   local no replica. Fix: GRANT ALL explicito sobre estas 2 en la 000.

2. "movimientos_inventario_tipo_check" violado (crear_devolucion,
   crear_recepcion). Mismo patron que el CHECK de series_numeracion de la sesion
   pasada: la 000 solo permitia ('entrada','salida','ajuste') pero produccion
   tiene 5 valores (+'ingreso','egreso', agregados a mano). Verificado contra
   pg_get_constraintdef real. Corregido en la 000.

3. "function is(numeric, integer, unknown) does not exist" en
   crear_venta_efectos_colaterales, caso 6c. comprobante_items.cantidad_entregada
   es numeric; el literal 4 se interpretaba como integer y pgTAP no puede resolver
   el overload de is() con tipos distintos. Fix: cast 4::numeric.

4. iibb_liquidacion fallaba con "Impuestos Avanzados no esta activado" — al
   fixture le faltaba `usa_impuestos_avanzados = true` en el INSERT de empresas.

Pendiente para la proxima: cheques_contabilizacion_166 tiene 4 fallos de ASSERT
reales (no excepciones/infra) — logica a revisar, no permiso ni schema.

+++ supabase/tests/crear_venta_efectos_colaterales.test.sql
@@ -264,7 +264,7 @@ SELECT is(
   (SELECT ci.cantidad_entregada FROM public.comprobante_items ci
    JOIN public.comprobantes c ON c.id = ci.comprobante_id
    WHERE c.numero_venta = 'V-CASO6'),
-  4,
+  4::numeric,
   'Caso 6c: comprobante_items.cantidad_entregada tambien se completa al reconciliar con entrega manual'
 );
